Adding Histogram Utility (#39)
* Fixing /result/latest and auth bug in diagnosis

* Bug check

* Bug check revertion

* Edited return error to throw error

* Skip test for make test

* Adding histogram data helper

+++ src/main/java/com/qwerty/cogbench/model/Result.java
@@ -34,10 +34,10 @@ public class Result extends Auditable<String> {
 
   @Getter
   @Setter
-  private Float accuracy;
+  private Double accuracy;
 
   @Getter
   @Setter
-  private Float time;
+  private Double time;
 
 }
