+++ src/test/java/com/qwerty/cogbench/integration/ResultControllerTest.java
@@ -81,8 +81,8 @@ public class ResultControllerTest {
     this.doctor.setRole("ROLE_DOCTOR");
 
     this.result = new Result();
-    this.result.setAccuracy((float) 78.9);
-    this.result.setTime((float) 78.1);
+    this.result.setAccuracy(78.9);
+    this.result.setTime(78.1);
   }
 
   public Result getPersistentResult() {
